A5-Naive_RAG_vs_Contextual_Retrieval/app/backend/requirements.txt
Django
djangorestframework
django-cors-headers
transformers
safetensors
torch
